Refactoring completed
