+++ rationale/Decentralized Betting (2).png:Zone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=https://lucid.app/
+HostUrl=https://pdf-service-lucidchart-com.s3.amazonaws.com/5feedd2a-7ca6-4975-b02f-621e569a82cc?X-Amz-Security-Token=IQoJb3JpZ2luX2VjENv%2F%2F%2F%2F%2F%2F%2F%2F%2F%2FwEaCXVzLWVhc3QtMSJIMEYCIQCyflSRzQdv28zmev42chfjbJ6mdOVqUdGg%2B7DYRWJV1AIhAMPsM1vblyvx%2FjDFoZpKxfkjo9%2BcSniO0n2YOIQHcGilKtUECPT%2F%2F%2F%2F%2F%2F%2F%2F%2F%2FwEQABoMOTM1NjA2OTI2MTgyIgwOqEVSk9LbEO3ZMjMqqQRL2yryHxYBzjtK8vE2R%2Fvq6NTTIg%2BF7rnFOy2RnUj0pSkBj%2ByolR%2FkIfAjqTVU%2BUME%2FwnXFb6Zc1RiNPpzBCK9Gwc%2Foc4Y0j8GzCzVKt9HZeR0ONtSwouFxnbeuJLsrHNg81ceMRuZ%2BbdRNhwhimLspuHxB79X8MZWlumc1V30wlXOswUXNDn68uRqm%2FHAJetxZZ1KrOpvkVtRDN3pxK4iLXcylWKVuw%2FhObB3vRAvTRlC%2BrjoNySNgtt0l77tl3XnQw8qj6iyRbw2dNdUQwsmVpc%2F3GiF7FchxAtaTn8vYo9W%2B8iVrxSjz%2BE%2FAl3f54aNK3Of%2BC9XF4pLRSGG%2Fh50aj4xPZ%2FZpMoaeSjl5Au%2FKVJX1c3sFwJ%2BMzKJN4CDCq5WVckn%2BT9FMcLO0ONyFXUTczy8GLOFlLiXvdc7cn0sDC8c9%2FTO8AQwGUAlvUBJUBXJ%2FSOEBDfQqd2JJMtc%2FWyRvfd%2FA0PnhXn4sonzVAGct7QZ1PLAzbAbb3gQAmdAeIllmz7O9K14pa8UHEx2eEyKBr%2FvfR4s3CffcrY%2FBS3Vc8dISznu8ask6WHF%2FYPppTd60%2F69kF6DyTy9YDQ%2F7SY4JL1Lsvs0AWTScPDbPjyoCq5unFhsMOPvyV29AZcUZDWRUtmVff6zWXfPIXaItj3pCfcuxGoBscTOGtFEQjm2JBSOHAp0%2BvSLP27D5N65nhAZMkomh5MUFR1gY0TSJLquwCxMVYHFi3SOMMj9uJwGOqgBbp%2B3mn2jOZRab0cn59vZdWb0HKs7%2F2eLxrrU90kj9%2BVoQp2g3M%2Fr43YIRzw1O0kB%2BqbOVvD7I5Tg%2BO%2BARhonsGwdx%2Fa59Fk%2BzTY1yfI7wEgER7ZqklX32ifaFAFeiQcANflBiF%2FChUeZJx3bJxCjstgXURkL7Ty97pbsyxei9hYKV7p84pbyzkqHp8lvpOqt%2FhXNZN1a6sizuUziILHjHOIv5Ot76RoZ&X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Date=20221205T194921Z&X-Amz-SignedHeaders=host&X-Amz-Expires=86400&X-Amz-Credential=ASIA5TVUEXNTOTN373QK%2F20221205%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Signature=abe385fbda2e7b299113ec725850e448be23f4e790102f7fd39b36f1d274cbfa
