--- src/Test.hs
@@ -1,10 +1,0 @@
-
-module Test where
-f1 :: Integer -> Integer -> Bool
-f1 x y = x >= y
-
-f2 :: Integer -> Integer -> Bool
-f2 x y = x <= y
-
-x = [1,2,3,5,7,8,5,1]
-
